Simplify status conversions
The current code used to convert the status string (chk_st, dm_st) to
status code when creating an instance, and provide a static method to
convert the status code back to status string.

Simplify by adding a status_string property returning the actual status
string wet got from multipathd, and reverse the _STATUS_CONV table so we
can use it to convert the status string to status code.

This make user code nicer:

   print(p.status_string)

Instead of exposing the internals of the module:

   print(DMMP_path.status_to_str(p.status))

This changes the status string to report the value we got from
multipathd, instead of normalizing it to "undef". This is more
consistent with the other properties.